fixes isTargeted remaining true
I had two functions trying to do the same thing and addRocketControls() was setting the property isTargeted of ships to true even before a rocket was fired, so new rockets just went by.

This commit entirely removes unused functions such as fireRocket() and  addRocketControls() and leaves all the targeting to fireRocketsForHand() which is now called by handlePlayHand() each time a hand is played.

Additionally, all targets are cleared at the end of the turn just as a safeguard.